+++ src/main/java/beyond/momentours/member/command/application/controller/MemberController.java
@@ -5,10 +5,10 @@ import beyond.momentours.member.command.application.dto.LoginDTO;
 import beyond.momentours.member.command.application.dto.MemberDTO;
 import beyond.momentours.member.command.application.mapper.MemberConverter;
 import beyond.momentours.member.command.application.service.MemberService;
-import beyond.momentours.member.domain.vo.reponse.ResponseLoginMemberVO;
-import beyond.momentours.member.domain.vo.reponse.ResponseSignupMemberVO;
-import beyond.momentours.member.domain.vo.request.RequestLoginMemberVO;
-import beyond.momentours.member.domain.vo.request.RequestSignupMemberVO;
+import beyond.momentours.member.command.domain.vo.reponse.ResponseLoginMemberVO;
+import beyond.momentours.member.command.domain.vo.reponse.ResponseSignupMemberVO;
+import beyond.momentours.member.command.domain.vo.request.RequestLoginMemberVO;
+import beyond.momentours.member.command.domain.vo.request.RequestSignupMemberVO;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
