+++ src/main/java/beyond/momentours/member/command/application/service/MemberServiceImpl.java
@@ -4,11 +4,10 @@ import beyond.momentours.member.command.application.dto.CustomUserDetails;
 import beyond.momentours.member.command.application.dto.LoginDTO;
 import beyond.momentours.member.command.application.dto.MemberDTO;
 import beyond.momentours.member.command.application.mapper.MemberConverter;
-import beyond.momentours.member.domain.aggregate.entity.Member;
-import beyond.momentours.member.domain.repository.MemberRepository;
+import beyond.momentours.member.command.domain.aggregate.entity.Member;
+import beyond.momentours.member.command.domain.repository.MemberRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
